Preserve fractional day and time fields in relativedelta multiplication
relativedelta.__mul__ wrapped every field in int(), so multiplying a
relativedelta that had a fractional day or time component truncated it
(for example relativedelta(days=1.5) * 1 became days=1). Drop the int()
coercion for the fields the constructor already allows to be fractional
(days, hours, minutes, seconds, microseconds); years and months stay
integral, as the constructor requires.

Fixes #1318

+++ src/dateutil/relativedelta.py
@@ -498,22 +498,24 @@ class relativedelta(object):
         except TypeError:
             return NotImplemented
 
-        return self.__class__(years=int(self.years * f),
-                             months=int(self.months * f),
-                             days=int(self.days * f),
-                             hours=int(self.hours * f),
-                             minutes=int(self.minutes * f),
-                             seconds=int(self.seconds * f),
-                             microseconds=int(self.microseconds * f),
-                             leapdays=self.leapdays,
-                             year=self.year,
-                             month=self.month,
-                             day=self.day,
-                             weekday=self.weekday,
-                             hour=self.hour,
-                             minute=self.minute,
-                             second=self.second,
-                             microsecond=self.microsecond)
+        return self.__class__(
+            years=int(self.years * f),
+            months=int(self.months * f),
+            days=self.days * f,
+            hours=self.hours * f,
+            minutes=self.minutes * f,
+            seconds=self.seconds * f,
+            microseconds=self.microseconds * f,
+            leapdays=self.leapdays,
+            year=self.year,
+            month=self.month,
+            day=self.day,
+            weekday=self.weekday,
+            hour=self.hour,
+            minute=self.minute,
+            second=self.second,
+            microsecond=self.microsecond,
+        )
 
     __rmul__ = __mul__
